--- core-java-module/src/main/java/programs/math/calculator.java
@@ -1,36 +1,0 @@
-package programs.math;
-
-public class calculator {
-    private int x;
-    private int y;
-    private int z;
-
-
-    public  calculator(int x, int y){
-        this.x=x;
-        this.y=y;
-    }
-
-    public void add(){
-        System.out.println(x+y);
-    }
-    public void subtract(){
-        System.out.println(x-y);
-    }
-    public void division(){
-        System.out.println(x/y);
-    }
-    public void multiply(){
-        System.out.println(x*y);
-    }
-
-    public int addr(){
-        z=x+y;
-        return z;
-    }
-    public int  subtractr(){
-        z=x-y;
-        return z;
-    }
-}
-
